pyproject.toml
[project]
name = "lzw-compression"
version = "0.1.0"
dependencies = [
    "Pillow",
    "numpy",
]

[tool.ruff]
line-length = 100
target-version = "py312"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP"]
